stripe integration done, need make documentation

+++ src/api/migrations/20230320091810-create-items.js
@@ -13,7 +13,7 @@ module.exports = {
 				type: Sequelize.STRING,
 			},
 			price: {
-				type: Sequelize.FLOAT(10, 2),
+				type: Sequelize.INTEGER,
 			},
 			createdAt: {
 				allowNull: false,
