+++ src/api/migrations/20230322191815-create-orders.js
@@ -22,7 +22,7 @@ module.exports = {
 				type: Sequelize.INTEGER,
 			},
 			totalPrice: {
-				type: Sequelize.FLOAT(10, 2),
+				type: Sequelize.INTEGER,
 			},
 			userId: {
 				type: Sequelize.INTEGER,
